Move jetsoncamera package
